Do not split live intervals for registers used for locking.
When we can avoid splitting live intervals for registers used for
locking we should always do so. Using spilling for registers used
for locking confuses the Art lock verification code and leads
to warnings (and those warnings are treated as errors by
certain builds).

With this change, the register allocator only splits live ranges
holding objects used in locking operations as a last resort. If you
have deeply nested synchronized blocks we might have to spill the
value in any case and we will. But for most code this should never
happen.

R=mikaelpeltier@google.com, sgjesse@google.com

Bug: b/64765186
Change-Id: Ic8d4519cff28f8720830a8a884d98a72c3f108e6