Improve the switch rewriting
For switch instructions with one or two outlier cases, but otherwise
consecutive cases, the switch is rewritten to a switch with only
consecutive cases preceeded by one or two if instructions checking the
outlier cases.

This change adds some simple support for estimating the size of the
final dex size of some instructions.

Change-Id: I3a1fea7bb249fa3cc8cd905acf4e4245d8c11b9b