Process inlining of methods in waves instead of ruling out breakers.
We used to never inline across an edge in the call-graph that was
removed during cycle removal to avoid non-determinism. With this
change, we only disallow inlining of methods that are being processed
concurrently. This is equivalent, as a method that is being inlined
concurrently must have had an edge in the call-graph that was removed if
it is called from another method. However, we do not rule out inlining
of such breakers across unrelated edges.

Bug:
Change-Id: I463fe850e677686624104060f48c2b2eab401c98