Fix const type of null constants from int to object.
Various optimizations consider constants the same if they have the
same type and value. Having null be of type int makes it possible
for the optimizations to collaps a const 0 instruction of type
int with a const 0 instruction that represents null. Correctly
marking null constants as having type object fixes the issue.

R=sgjesse@google.com

Bug: 64740479
Change-Id: I16a3cddda673865f1ddc57e1dbf0a422dc912e27