Instructions can be targeted by both normal and exceptional flow.
This CL fixes the state to ensure that exceptional edges have a single element
on the stack and that normal flow can target the same instruction. At build time
all exceptional edges are directed via a header block with the dex move-exception.

This fixes bug 65060345 where we failed to merge the stacks of an exceptional
edge and a normal flow edge. The fix exposes a non-deterministic failure in R8
for the same test.

Bug: 65060345
Bug: 65233869

Change-Id: Ifc2f8838bcde588fed96c58e4f92a9929631560c