Avoid duplicate debug positions for exception targets.
In the common case where a block is only targeted by exceptional
edges, we get duplicate debug positions at the move exception
instruction and at the actual target. That typically leads
to uneccessary nop instructions in the output.

This change removes these duplicate debug positions when it
is safe to do so.

R=zerny@google.com

Change-Id: Id284d8e3664b76f7a4848c4d23d08f8b81533c1e