Use ART default options
For each version of ART runtime, we want to test the default
configuration that is executed on real Android devices. Therefore we
should rely on default command-line of ART to execute a test.

This CL removes explicit configuration options to rely on default
configuration. In the future, we might want to test multiple
configurations of a runtime as well.

Note: debugger-related tests still need to pass the --debuggable
argument in order to execute test with full debuggability.

Bug: 65004695
Change-Id: Ic8885166a6fe631b27af1763d94f0ce48797cd38