Allow const number instructions to define locals directly.
By allowing the collapsing of:

v0 <- ConstNumber 0
v1(x) <- DebugLocalWrite v0

to

v1(x) <- ConstNumber 0

when the debug local write is the only user and there are no
positions between the definition and the debug local write, we
avoid the register allocator chosing registers that force the
materialization of the local write as a move.

In order to make this work, we need to make sure that constants
are not moved if they directly define a local. Shortening of
live ranges has been updated with that. The rest of the code
should already be careful about this.

R=zerny@google.com

Bug: 66161160
Change-Id: Ie0d46375c13cbfb71f08c7a39eb26412b1d76c28