Fix download dependencies when output folder does not exist
The "download_deps_..." tasks only considers the .tar.gz file as an
output. If the dependency folder does not exist, the archive file is
not uncompressed. This CL fixes this by adding the output folder as
an output as well.

Also rework how dependencies are listed by removing the .tar.gz
extension which is common to all dependencies. Therefore dependencies
are now declared by their output folder.

Change-Id: I22c5b82887d1ce2070a558c0ca24cbe40157c499