Proposed minimal DiagnosticsHandler API.
Add a method to D8Command and R8Command to supply a
Diagnostics handler that will be used for all diagnostic information
from the compilers.

This is currently only used for CompilationException and the warnings
generated in connection with default interface method desugaring.
Once we agree on the interface, we should make sure that we use this
instead of stdout and stderr for all warning and info diagnostic.

The current default implementation of the API keeps the behavior of
the current code. However, the API allows us to extend the Diagnostic
messages with more information over time. Also, with errors passed
to the DiagnosticsHandler we can consider making certain errors
non-fatal and keep on trucking to get more errors reported in one
compilation.

R=benoitlamarche@google.com, gavra@google.com, zerny@google.com

Change-Id: Iedd0b7d71460c14a78c13c09b340e491071c2fae
Bug: 67087703