Enable run of tests on devices and emulators
This CL updates the DexVm enum to add add an information
about the kind of runtime used to run the tests, whether
it's a host version of the runtime or a target one
(device or emulator).
It adds also the DeviceRunner class which drives the target
device for the tests.

Change-Id: I10666951f083a4f1da2efe3e9d368803a223a1b2
Bug: