Remove error method from DiagnosticsHandler.
Since the error method has a checked exception it will be really
annoying to use internally in the system. Also, at this point
we never produce more than one error before we terminate
compilation.

We need to think more about how to handle this. For now it seems
like a good start to just deal with warning and info in the
DiagnosticsHandler so that users can decide where to put that
information instead of us spamming stdout/stderr.

R=benoitlamarche@google.com, jplesot@google.com, yroussel@google.com

Change-Id: Ic4c7eca271e7671dbb092c33f2b07b89ed9403ae