Workaround flaky Dalvik shutdown issue in debug tests
When we resume the Dalvik runtime (using the VM.Resume command)
during debug tests, it may shutdown without sending the reply to the
command, leading the test to hang until a timeout. This is causing a
flaky failure with the TestErrorException being thrown.

This CL works around the synchronization issue by ignoring this
exception if and only if the resume was the last command of the test.

Bug: 65869324
Change-Id: I4ca963e121ace4fc1b8bda81b76dafc11631171f