Fix register hints code to work when register is revoked.
In certain cases the register allocator can decide to revoke a
register that was assigned to an interval. The register is
optimistically assigned but turns out to be needed for something
else. If the interval for which the register is revoked is
the hint for another interval, the current code will lead
to an ArrayIndexOutOfBounds exception. This change fixes
that by correctly treating a hint of NO_REGISTER as if
there was no hint.

R=sgjesse@google.com

Change-Id: I7a18e295a329d65867fac553d1a259fe81c2b217