Ensure member arrays are accessed synchronized during write.
Now that we share classes between two files, we have to make sure
that we do not access the member arrays while updating them. The
referenced bug was caused (I believe) by us using the array to
collect items in one thread while another thread was sorting them.

This CL changes synchronization to be per member array instead of
per dex class. This is mostly to make it more visible what we are
synchronizing on.

Bug: 67258314
Change-Id: I28a9bb68335673377032e0470ca5aef4c96e87ca