Support skipping of class loader frames in debug tests
The default debug filtering of IntelliJ skips frames in class loader.
This CL adds the same capability to our debug test infrastructure.

We used to skip frames by repeating the same step command. But to
skip class loading, we actually need to step out of the current frame
so that the whole class loading code has executed. In order to achieve
this, this CL refactors how step filtering is implemented. Now the
StepFilter instance is reponsible for adding extra step commands into
the commands queue when it is necessary. By extension, the 'step
until' capability is also a StepFilter instance now.

Since we can skip class loader frames now, this CL also enables the
tests that were skipped for Dalvik due to class loader frames:
* com.android.tools.r8.debug.InterfaceMethodTest#testDefaultMethod
* com.android.tools.r8.debug.MinificationTest#testBreakInMainClass

Bug: 67225390
Change-Id: I112161be31cc99ddcf1eeaa1b6a63a97a311b4d6