Fix bug in inactive linked interval splitting.
Linked intervals have fixed registers for all uses to make sure
that values are in consecutive registers at ranged invokes.
Therefore, when splitting an inactive linked interval, we need
to split before the next use and make sure that the same
register is used again at that use. However, if there are
no uses after the split point, there is no need to split and
attempting to do so will lead to an index out of bounds
exception.

R=sgjesse@google.com

Bug: 68656641
Change-Id: Iacd878895532f80c24adb52e9ce4c9abed57bf98