Use more lit8 dex instruction form
- During build of lit8/lit16 dex instructions what we want to
check is if a value need a register for the current instruction
or not. Check if the value need a register for all usages is too
restrictive, for instance, each time a constant through by a phi
instruction, lit8/lit16 instructions were not used but it is safe
to use them.

Bug: 68188667
Change-Id: I50bc781c3ecde0fa5556355dab263644509a9c39