Add fix-point calculation to the enqueuer for instantiated classes
The set of instantiated classes are build during enqueuing. Checks
dependent on the set of instantiated methods need to be handled in a
fix-point fashion.

Items rejected due to classes not being instantated, are now tracked. After
each tracing iteration these items are checked again if new instantiated
classes have been discovered in that iteration.

Bug: 68696348
Change-Id: I41410dae85a031aec1adc3f6b8a5883de9017042