Special handling of referenced classes in Proguard compatibility mode
Keep classes and their default constructor for classes referenced when running
with --force-proguard-compatibility for the compat proguard command if the
class actually does have a default constructor.

This covers the special handling of const-class and include classes used in
check-cast.

For now the keep reason will be "invoked by the default constructor", which is
somewhat misleading, but until we know to what xtend this is actually needed
this was the simpel solution.

Bug: 68754879
Bug: 68689582
Change-Id: I6761fcd50156b19248c080ef8e57e43bcc570f15
Bug: 68754879