[CF] Compute stack-map frames.
This adds stack-maps for all locals. This is very simple at the moment because
stack is empty on all joins of more than one control edge, thus only types of
locals are needed. Due to the lack of register allocation, local indexes are
invariant and map to a unique SSA value or phi. The verification types are
computed as a fixed-point over the flow of values (so phis will resolve to
the join/least-upper-bound of its inflows).

Change-Id: I300f41c86e9a71d0df83944c7f9aee08f35f0ba9