Refactor debug-test configuration out of the debug test runner.
This is a first step in cleaning up the debug tests which have gradually been
extended with compilation routines and configurations that should rather be
specified on individual tests and separate from the actual running of the test.
In particular, the test configurations no longer have static references into the
temporary directories of the individual tests. Follow-up changes will remove the
remaining compilation routines from DebugTestBase.

Change-Id: Ieef6ca1ea067a54c8539ebb3e1f9eec0b49137dc