Do not lower const-number instructions past array-length.
This is a workaround for an invalid bounds-check elimination bug
in certain Lollipop Art versions. Those Art versions generate
bailout code based on the array length and jumps to a slow version
of the code where the loop variable might not have been initialized.
That leads to ArrayIndexOutOfBounds exceptions. This speculatively
attempts to put the const-number instruction that initializes the
loop variable before the array-length instruction in an attempt
to make these art VMs initialize the loop variable before the
bailout check.

There are no guarantees that this works at all. However, the size
impact is small and it works on the reproduction from okhttp.

R=sgjesse@google.com, zerny@google.com

Bug: 69364976
Change-Id: Iaee239395f6fa8740e7368cde0b94f349ae07fc6