Always set ACC_ABSTRACT and interfaces for DEX.
Some older javac versions produce interfaces that are not abstract for
package-info files. We accept those and read them just fine, as for
class files we can gate on the version of the compiler that generated
them. For DEX files, this is not true. So, instead, we now always add
the missing ACC_ABSTRACT flag.

Bug:
Change-Id: Ic2c7acf245d1c47f59c336c4671386162425e5fe