Prepare to make ProgramResource part of the public API.
This moves the previous internal ProgramResource into the public visible ns. The program
resource is intended to replace the access to class-descriptors on bare resources, so the
resource is meaningful for all application resources (byte/file info for code, pg-conf,
pg-map etc.). The deprecated methods have been marked as such for now so we can iterate on
the interface and do the API breaking removal at a later point.

Change-Id: I0eec0eff642cd4aea149ea8d380d07714a8a2372