Fix return type adjustment for method references (b/70348575)
Java provides more flexibility for signatures of samMethod, implMethod and
instantiatedMethod for method references comparing to regular lambdas.

In particular, implMethod return type does not have to be a subtype of
samMethod or instantiatedMethod return types.

We have to add an extra cast to avoid verification error in such cases.

Note that since we don't have whole class hierarchy, we have to be
pragmatic and always add cast in case these types are both reference
types and are different.

This may result in few unnesessary casts, but guarantees correctness.
Extra casts may be removed later with optimization like b/70026120.

Bug:70348575
Change-Id: I696ebd21c32089c07d8210566314db48e25ec679