Introduce dedicated dex-file consumers.
The consumers for DEX files are split into two. One for indexed mode and one for
file-per-class-file mode. The type of consumer defines what output-mode
compilation is in. The explicit output-mode and output-path getters and setters
are deprecated and will be removed once the API supports setting up the
consumers directly.

Change-Id: I58ade03d221f1571a8db79f7f7bc573f8a479c8b