Introduce program consumers in the public API and deprecate setOutputPath and setOutputMode.
The configuration of output can now be done in one of two ways:

1. Set output path and mode using setOutputPath and setOutputMode on the builder.

2. Set a program consumer by using setProgramConsumer or setOutput(Path, OutputMode). In
   both cases a program consumer is defined and will be used for receiveing output.

For R8 only API 2 is supported. The R8 command-line has been updated to allow the
mutually exclusive flags --dex and --classfile denoting the output format.
The implicit default is --dex.

For D8 both 1 and 2 are supported now, but they are mutually exclusive. The builder will
abort if a mix of the two appear.

For D8 we retain a backwards-compatible default set-up using the deprecated API, such that a
default builder, eg, D8Command.builder().build() will an empty dex-indexed consumer. This
behaviour is deprecated and for R8 will result in an error.

Change-Id: I3b387d4bcd6b0921e7e6f46b22f6d8244c7132b0