Make SparseConditionalConstantPropagation resilient to NPE.
While working on topological sorting of basic blocks, found that
Instruction#evalute may raise NPE if either operand's lattice is not
computed yet. That is, depending on basic block order, the analysis may
or may not encounter NPE. One conservative yet effective way to avoid it
is to return TOP if there is no lattice bound to the given value.

To do so, instead of explicitly passing (sort of) internal mapping, we
can abstract the map from value to lattice, and allow evalute() to rely
on that map's get(), where we will return either bound lattice or TOP.

Change-Id: If05681ae8be21e865c6e169ada88e25d0b429090