Allow meaningless phis during dead block removal.
Additionally, remove more trivial phis. When inlining and after
constant propagation we were leaving trivial phis behind. This
can get in the way for further optimizations and we should always
get rid of trivial phis in the IR.

Added assert that there are no trivial phis when we expect the
graph to be consistent.

R=sgjesse@google.com, zerny@google.com

Bug: 70736958
Change-Id: I2ed418568d18e57230343b8f55ce994336d379a8