Fix implementation for invoke super.
The tree shaker and inlining will now take the calling context into
account. Furthermore, we now also handle the case of invoke-super
on a private methods, which may arise from translating invokespecial
in the jar frontend when calling a private method from the wrong
context. The effect of the rewrite will be the same.

Bug: 70707023
Change-Id: Ie0fb3cc84b3af50409e5656de9875c696a9c7009