Revert adding markNeverNull to InvokeNewArray.
The way InvokeNewArray is generated/refined by IRBuilder/CodeRewriter
made it hard to mark the out value neverNull. It sometimes generates the
IR with null out value, which causes NPE. It sometimes redefines the IR
while reusing the same out value, which causes assertion failure:
markNeverNull expects to be called only once in a value's lifetime.
Since this mechanism would be replaced by local type/nullability analysis
anytime soon, and not having that so far has not made any issues so far,
I'd simply revert that.

Change-Id: Idf1a15e84a74da50ec3a79177efb4dce132901ce