Keep InnerClasses and EnclosingMethod together in Proguard compatibility mode
R8 require that the attributes InnerClasses and EnclosingMethod are kept
together. In Proguard compatibility more make one infer the other to avoid
R8 compilation error requiring changes to the Progurad configuration file.

A -keepattributes directive with both attribute names is added to the
generated Proguard compatibility configuration in this case.

Bug: 69445518
Change-Id: Ie1514d274e8e9a2a467d4b8f803fed9ff95417e3