Make SwitchMapRemoval state explicit in AppInfo.
This is to get rid of the concept of extensions. Also, enable removal of
switchMap classes in debug mode, as well. After inspecting the code, I
cannot see where it should go wrong. It is very defensive anyway, so if
it fails, it will simply leave some maps in the code.

Bug: 62936642
Change-Id: Ie76c4ba10d4c33253368fb58b5b43a84acfb6f32