Allow inlining of methods checking null receiver
We used to only inline methods where the receiver is known to be
non-null. Otherwise an unused null receiver would not throw the
expected NullPointerException.

This CL extends inlining to methods that check the receiver (using
instructions that would throw if receiver is null) before any
instruction with side effect. Therefore, if the receiver is null,
the inlined code would throw the expected NullPointerException as
well.

The first step is to mark methods doing such null check in the
CodeRewriter phase. This is done before the inlining phase (and only
when inlining is enabled).
The second step is to let the inliner know that the candidate method
has the "null receiver check" flag to decide if the method can be
inlined. If that is not the case, then the candidate could be inlined
if the receiver is known to be non-null (among other inlining
constraints).

Bug: 70572176
Change-Id: Ia483d7536f957629c1ee01be08887ffbe932643e