Add synthetic keep rules when disabling the tree shaker.
As the tree shaker also is used as our main global analysis vehice, we
have to run it over all code that we will emit in the end. Hence, if we
disable tree shaking, we have to add rules to run it over all code.

Also keep all attributes in this case.

Bug: 72485384
Change-Id: I32440a8dbc04e522bea62b8d99834672dd4fc39c