Remove warnings from main dex calculation
Don't even warn about library classes extending program classes for main
dex calculation. It is quite difficult for the user to do much about these
errors - especially in the Android Studio Setting, as it require adding
filters to the input for the main dex list compilation.

Add a test which exercises exactly this JUnit setup.

This also adds support for a diagnostics handler for the main dex list
and Proguard compatiblity command.

Bug: 72312389
Change-Id: I7aba5275c48ce37242547191750dda10896b00ae