Flipping if statement in case fall-through successor always throws.
In case an if instruction has fall-through successor which always throws an
exception while the other ('true-target') successor does not, it is very
likely that this if instruction serves as a safeguard for assertion-like
test, for example for null checks.

In this case we assume that always throwing branch is executed very rarely
and 'true-target' branch is what is executed more frequently since it
represents no error scenario.

For such if instructions we flip the branches and invert the test, so the
non-error branch becomes a fall-through successor. This helps improve
performance on ART.

Bug:
Change-Id: I814003fe4487ca3a55c1acb77b5b4d11fc310b3c