Add tests for current line-table issues.
The main issue is that local-only changes must use only advance-pc events. Doing
so will cause issues with break points in finally blocks, but this is an
existing issue with JavaC compiled code as verified by the tests here.  These
test are currently failing for D8 compiled outputs. Fixes will be done in
follow-up CLs.

Bug: 72933440
Change-Id: I3e86b2c55398aa8d57592ab221bae53d99bb8814