Flipping if statement in case fall-through successor always throws (#2).
In case an if instruction has fall-through successor which always throws an
exception while the other ('true-target') successor does not, it is very
likely that this if instruction serves as a safeguard for assertion-like
test, for example for null checks.

In this case we assume that always throwing branch is executed very rarely
and 'true-target' branch is what is executed more frequently since it
represents no error scenario.

For such if instructions we flip the branches and invert the test, so the
non-error branch becomes a fall-through successor. This helps improve
performance on ART.

Note that we also do NOT flip the successors if the method code has any
positions and any of throwing instructions in fall-through successor
does not have a position. This should guarantee we don't introduce wrong
debug information by moving such an instruction to wrong place.

Bug:
Change-Id: I1243c1ff8e092c39211a0a8228ad23c76448e06c