Add the chosen field renaming back to NamingState.
NamingState#assignNewNameFor() litereally assigns a new name for given
proto. It is up to Minifier which new name to use, though.

For example, MethodMinifier computes the frontier map of interface
hierarchy and tries to assign and reuse a unique name that avoids
conflicts among interfaces. To do so, for each state that corresponds to
different interface, it keeps trying to assign a new name until the new
name is available at all interfaces. At the end, only the chosen, unique
name is stored at the top-level state to reuse (for the same name) or
avoid conflicts.

FieldMinifier misbehaves for -useuniqueclassmembernames: it should
assign the same rename to the same name if any. The bug was, it never
stored which name was chosen in the NamingState. (Thus, always returns
new names.)

Bug: 72858955, 73149686
Change-Id: Ia293d47e9e2e4dcc07fc46a1f59cc02637b9a5b3