Resolve field name conflicts in aggressively overloaded input.
MemberMinifier has two-layered lookup:
1) from proto (either field type or method signature) to internal state,
2) inside internal state, original name to newly assigned (re)name.

-overloadaggressively tweak what to use as a key at the 1st layer.
If enabled, field type is always mapped to void, and thus fields with
different type can be named to the same rename.
(For methods, return type is ignored.)

-useuniqueclassmembernames determines whether to remember name mapping
inside the internal state, i.e., entry in the 2nd layer.

------

This CL clearly settled the renaming behavior regarding the 2nd layer.
With -useuniqueclassmembernames, we should keep renaming mappings in an
internal state. On the other hand, if it's not set, by not maintaining
such renaming, we can assign different names to already aggressively
overloaded fields.

------

Method renaming has a different story, since overloading methods in
subtypes should use the same renamed names. Will handle that in a
separate CL.

Bug: 73149686
Change-Id: I79ce71b4da485bdb4e5523e60f149e39e3c1d6a9