Relax the inlining constraints for virtual invokes.
We checked the access flags of the containing class of all the
potential targets. That is not necessary. We need to check the
access flags of the containing class of the resolution target.
For the actual targets we only have to check the access flags
on the target method itself, not of the containing class.

BUG: 73738623
Change-Id: I5b499abc3e3feaa03dd964b8b7ff8d2681eb8f66