Devirtualization from invoke-interface to invoke-virtual.
If we can nail down the type of the receiver, we can replace interface
call with devirtualized, virtual call. E.g.,

  invoke-interface i, ... I#foo // i is of type A

  ~>

  a <- checkcast A i  // Otherwise, ART verification error.
  invoke-virtual a, ... A#foo

Here is the dilemma. We can't replace all the occurrences of i with a
because the uses of i are unclear. On the other hand, we can't insert
that additional cast every time we devirtualize interface call.
Otherwise we will see something like:

  a1 <- check-cast A i
  invoke-virtual a1, ... A#m1 (from I#m1)
  ...
  a2 <- check-cast A i  // We can reuse a1!
  invoke-virtual a2, ... A#m2 (from I#m2)

Hence, cache'ing:
  original receiver -> holder of devirtualized call -> casted receiver
(The same receiver can be used as different types of casted receivers.)

Bug: 69962188
Change-Id: Ib02280b7f8a7b6e8101271fa2eab49b2f618077a