Make desugaring more tolerant on missing hierarchy
This is done by skiping default method desugaring of classes with
only unimplemented defaults methods defined in direct interfaces defined
in libraries. We were already not generating code for those default
methods but we were still checking the hierarchy even there was only
such defaults methods, This change is just removing this hierarchy check
and has no impact on generated code.

This change is not a big relaxation of constraints, it is allowing the
particular case like implementing Iterator but desugaring of classes
with missing hierarchy will still fail if any other default method is
defined outside libraries.

Bug: 73633268
Change-Id: I9dad556de15c939a13ddec13f7ce0fbfd24a0bc7