Inline static methods preserving class initialization
When a class has a non-trivial <clinit> method, we used to forbid
inlining of any of its static methods to make sure that the
initialization of the class is preserved.

This CL relaxes that constraint when we can prove that the method
does preserve class initialization before anything else. That is to
say that the method will unconditionally execute an instruction that
triggers class initialization before any other side effect happens.

This improvement covers Kotlin accessor methods in companion classes.
These static methods are used to access properties of companion
classes which involves accessing a private static field in an outer
class. This CL also adds new test cases to verify that these Kotlin
accessor methods are inlined as expected when access relaxation is
enabled.

Bug: 70158739
Bug: 73518823
Change-Id: I079c2c8227ca085ee3637a9499dbef614baebc6f