CfConstNumber: Handle negative zero correctly
When `value` is float or double, `value == 0` is true even for
negative zero. The FCONST_0 and DCONST_0 opcodes output positive zero,
so they must be followed by FNEG or DNEG in the negative zero case.

An alternate fix would be handling negative zero with a LDC; however,
this always requires 3 bytes of opcode for double (LDC2_W takes 16-bit
index) and would sometimes require 3 bytes for float as well (depending
whether LDC or LDC_W is emitted). OTOH, ?CONST_0 + ?NEG is only 2 bytes.

Change-Id: I447339a0644abf50f2c0c84fc6ebf031ba92d555