CfRegisterAllocator: Fix linear scan adding active register to freeRegisters
If an inactive interval overlaps with an active interval, then its register
is not in freeRegisters and must not be in freeRegisters. Previous code added
some inactive intervals to freeRegisters after handling the current interval
without checking if the inactive interval was free before.
Art123_compiler_regressions_mtTest failed because a register was assigned to
two overlapping locals (Java variable names s and v2 in the test).

Try free registers until a register is found for which no inactive interval
with that register overlaps the current interval.

Change-Id: Ifa380d87f397b1d093d5a0fdbab7cd62d0a6f62b