CfApplicationWriter: Add private bool flags for debugging
Since ASM's verifier is run on individual classes, it will generally
fail with ClassNotFoundException (since it uses Class.forName() to look
up classes when type checking). The verifier output is mainly for
debugging, so we don't want it spamming the output of tests by default.
Instead, hide it behind a private static final boolean flag.

Also add a flag for printing the CF data.

Change-Id: I350d6a94731967d9ee3413630ebde267add44a43