Use both name and proto as key in internal naming state.
Once the given proto is used to search for the bound internal state, it
used only *name* to distinguish members. It was safe for javac-generated
bytecode, but that's not the case for general bytecode in compliance with
JVM-spec: the same name could be reused as long as member signatures are
different.

Internal naming states are chained according to the class hierarchy, and
then used to lookup how members in the super type were renamed. For
already aggressively overloaded input, name-based lookup was not able to
resolve members in the super type with the same name. To handle bytecode
in compliance with JVM-spec, we should use both name and proto as key.

One caveat is -useuniqueclassmembernames. To map members with the same
name to the same renamed name, we should use only name as key.

Bug: 73149686
Change-Id: Idb761ad0fa62c8a5fa84407f80fac7d1371aa6da