Add resource shrinker API to R8
This is a temporary API to allow
extraction of integer and string
constants from dex, which is necessary
for resource shrinking. In addition to that,
accesses to static fields e.g. R.layout.my_layout,
and method invocations e.g. Resources.getIdentifier()
are passed through to the API.

This API might be removed at any point.

Test: ResourceShrinkerTest
BUG=74109805

Change-Id: I2d3d8caa31d30d33c3df6724b7cb7cf1b8c479d5