Keep fewer default constructors in Proguard compat mode
Only keep the default constructors for classes where Proguard is
known to do the same.

Before this change we unconditially kept default constructors
for all classes mentioned.

This also changed the semantics of how we handle -keepclassmembers, as
we don't include the default constructor when the class is only
mentioned. This cause the instance methods kept through
-keepclassmembers for a class which is not instantiated to become
abstract. Proguard actually keeps the code for these.

Bug: 74233021
Bug: 69028743
Change-Id: If3d85b6e21e55a48ac933c50a10e9eff7a6f737b