Make dexsplitter default to base
In the current form, when d8 does e.g., desugaring of lambdas, we might
get classes that would not have a mapping based on the jar files.

This will make the splitter push all classes that does not have a mapping
into a split called 'base'. You can still overwrite where non matched
classes should go with a *:split_name rule in a mapping file.

Change-Id: Ic68bd7ee3f9f388ccf7b10e4f8ba0d83d6529b66