Fix a wrong split in Devirtualizer#devirtualizeInvokeInterface.
If the currently visiting block has catch handlers, we splited blocks
to be a consistent SSA. However, that split was done after adding a
necessary checkcast. Since checkcast is also a throwing instruction,
catch handlers were bound to that newly checkcast, not the devirtualized
invoke.

This CL fixes the order: split, if needed, *before* adding checkcast.

Bug: 69962188
Change-Id: I842699548824249aa44137d3648ffbebebe2ace8