Show receiver in debug info of desugared default method
When we desugar a default method (of an interface), we move its code
to a static method of a companion class. In order to access the
original receiver (instance of the interface), we pass it in the
arguments of the method.

For debugging, we must give it a name so that it can be seen by the
debugger. Though its natural name would be "this", we cannot use
that name because the debugger may hide it (due to the method being
static).

In order to workaround that issue, we choose the name "-this". Indeed
it starts by a character that should not appear in the source code.

Bug: 62290295
Change-Id: I5064701cb22f6b84bea02ce5efe11fcf296968a9