Add OpenJDK 1.8 JRE to third_party
In CF tests that run the R8 output on the host JVM, we need to compile
the test program with the Java 8 library class hierarchy and not the
class hierarchy of an ancient android.jar.

In particular, the shaking12 test in TreeShakingTest fails with the CF
backend because InstantiationException and IllegalAccessException extend
Exception in Java 6 / API level 18 and lower, but they extend
ReflectiveOperationException in Java 7 / API 19 and higher. The test
fails because the input class, compiled with host Java 8, contains a
call to e.printStackTrace() in an exception handler where e is a
ReflectiveOperationException, but R8 emits this call in a context where
the verification type of e is Exception since R8 uses the lowest API
level available where the LUB of InstantiationException and
IllegalAccessException is Exception.

Change-Id: I659f8c0824d5154b77f460082fcb4444a39974a3