Support private static interface methods
In Java 9, InvokeStatic and InvokeSpecial take either a MethodRef or
InterfaceMethodRef depending on whether the type receiving the invoke is
a class or an interface.

Add 'itf' flag to IR instructions InvokeDirect, InvokeStatic and
InvokeSuper to track whether the method is an interface method.

Add 'itf' flag to CfInvoke and pass this flag to ASM when writing.

Fixes two Java 9 examples tests involving private interface methods.

Change-Id: I6f16f3e18c2b368a1e89d8e5f7d7d81dc41cb2b4