Fixing super calls to default interface methods in desugaring.
When interface or class calls default interface method via supercall
invokespecial instruction must point to an interface which is directly
implemented/extended by the class/interface performing the call. In
this case, invokespecial instruction may point to a default method
not directly implemented by the interface, but inherited by it.

We try to make our best attempt to find proper method to call, and
only use the method specified if we fail to do so.

Bug:
Change-Id: I52498aff4cf508bd4815dc4a3fac5ff24c7413af