CF backend: Compute uninitialized types in stackmaps
Add new type CfFrame.Uninitialized representing either an uninitialized
reference from a NewInstruction or from the `this` argument in an
instance initializer.

Construct a mapping from each NewInstance instruction to the Invokes
that initialize the reference, and construct the list of Invokes that
initialize the uninitialized `this`.

When emitting a stack frame involving values with definitions pointing
to NewInstruction or `this` Argument, follow predecessor pointers from
the current block to see if the reference is initialized at this point.

Since most code will have the NewInstance/Argument and the initializing
Invoke in the same block, there's no need to cache the calls to
findAllocator().

Also assert in CfBuilder that stack is empty between all basic blocks.

Change-Id: I949f951c9e7ec3a48162f0f8375faf0ba820d045