Add non-null IR _after_ inlining.
non-null IRs have been added when building IR code for inlinee. Since
blocks in the inlinee aren't linked to the caller yet, users of non-null
value in the caller were not linked either, i.e., not shown to inlinee,
and thus those uses were not properly updated or sometimes addition of
non-null IR itself has been skipped.

The key change in this CL is to run NonNullTracker *after* linking the
inlinee blocks. Also, found a number of small bugs in NonNullTracker and
corresponding tests.

Bug: 76200247
Change-Id: I02d56e6a5ae8e06da5956d969f24b6267ba4c1ea